Prelimary AOO VOO AAI VVI
This commit adds the necessary methods required for A1 however the simulation does not run if you don't comment out all the subsystems you're not using. This is because you can't reference the same pins in multiple subsystems. To correct for this I will make the individual subsystems output their signal into a merge block and hopefully get around the issue by using that way instead.