AOO, VOO, AAI, and VVI as one state flow diagram
This commit improves upon the preliminary 4 modes created in subsystems by merging them all into one state flow diagram which can have a "mode select" that chooses which mode gets iterated through when the simulation is run.

It looks like all of the modes are functional in the simulation but when exported over to the board there are weird behaviours that are noticed through the heartview program. The VOO mode does not pace the ventricles and instead paces both the atrium and the ventricle and the AAI and VVI modes can't seem to tell when the heart is pacing even though the sensitivities are properly set.

Proper implementation of hysteresis has still not been added to the simulink stateflow yet.