Adds hardware hiding
This commit has two big things in it. First is the hardware hiding. All of the inputs, stateflow logic and pins have now been grouped into their own subsystems. The subsystems then have inputs and outputs that allow them to be connected together within the simulink document.

The second addition contained within this commit is the addition of individual simulink models that were used to test AOO, VOO, AAI, and VVI. There was an issue where the board was not outputting the expected results from the simulink tests so I thought it could have stemmed from a niche error that could be rectified by dividing up the model into parts. Unfortunately that was not the case but I'll save these individual models such that we can keep them for small scale testing in the future.