Removes hysteresis. Adds fixed input names and minor improvements
Upon realizing that the variable names don't have to be the same as in the SRS VVI document, the names have been changed to be more readable. Hysteresis has been removed as that is not a requirement in A1. VVI was not working properly because the frontend control gate was not triggered, that was fixed.

While it is a minor improvement, the connections between subsystems have been changed so that the wires don't cross over each other as much.