Adds accelerometer signal processing, variable LRL, and smooth LRL changing
This commit is a big one.
It essentially provides the foundation for implementing AOOR, VOOR, AAIR, and VVIR without going the final step.

In this commit I've processed the accelerometer signal by filtering out the high frequency changes and by averaging out the acceleration over a period of 15 seconds to represent "average activity" from there the whether the average activity is between 5-6, 6-7, 7-8 or 8+ determines whether the heart beats at 60 bpm, 90 bpm, 120 bpm or 160 bpm.

From there the stateflow chart proves a smooth transition between the current LRL and the destination LRL by slowly changing it linearly within the timer period allowed.

Next steps are to fix the existing stateflow such that you can change modes during operation and provide a decision structure whether or not to override the input LRL with the accelerometer determined LRL.