Adds modified stateflow for dynamic mode change
This commit adds a new stateflow that will allow the DCM to change the mode of the pacemaker during it's operation as the mode is constantly evaluated throughout the stateflow progression.

Right now there is an issue with timings as the states seem to delay the timing by about 1 millisecond. I've posed this issue to the TAs and I'm waiting to hear back.