Adds untested reaction factor
This commit builds off the past series of commits by adding in the reaction factor to change how quickly the LRL will increase or decrease. The way I have it working is that each reaction factor corresponds to making the rate change 2% quicker from 0% quicker to 30% quicker.

This commit is untested though because shaking the board on the 6th floor of mills or during my lecture would be incredibly impolite. I'll test it later today at my earliest convenience.