Adds more separation of concerns
This commits puts the logical process in simulink using the function blocks into their appropriate subsystems, grouping them for easier organization and shows the software design principle of "separation of concerns" because each separate logical process is isolated from the other logical processes thus making it easier to debug and change without messing up other parts of the simulink logic.

Another change made was a slight change to the stateflow. I removed the collector state and move the transitions appropriately.