Adds simulink accelerometer demo and DDD implementation
This commit adds a scaled down version of the accelerometer subsystem with less LRL options so that it is easier to test the timings of the LRL changes as you won't need to hold an extremely precise activity to measure if you've abided by your reaction time.

This commit also adds the preliminary implementation of DDD and DDDR into the Group44PACEMAKER.slx file.

The timing issues from the other modes seem to reappear in the simulation and when deployed to the board there are major issues as the heartview output does not match the LRL while the simulation does.