Adds custom coded moving average function
This commit removed the low pass filter which didn't serve to do much to process the accelerometer signal and has a custom coded moving average function.

The simulink moving average block seemed to be so computationally heavy that it slowed down the processing of the FRDM board by about half. By removing this simulink function block and replacing it with a more efficiently coded function block, AOO was able to pace closer to its set LRL.

Going forward the accelerometer subsystem needs to be remodeled to support activity threshold and response factor with a curvilinear response.