Integrates serial communication with stateflow
Note: This commit is untested on the hardware

This commit integrates the serial communication test housed on the serial communications branch with the existing stateflow.

The simulink file has been changed such that the accelerometer also reads from the serial communication input.

Next step is to test the simulink file on the FDRM board.