Fixes major timing problems with pacemaker modes
This commit changes the parameters inside the accelerometer block and removes the simulink moving average block to attain a general timing delay by a factor of 2. This is better as we were originally at a factor of 6.

This commit also readded back the custom moving average block because it makes prototyping easier. The simulink function block slowed down my stateflow timing as a function of how big the window was. The custom function only slowed it down by a factor of 2 constantly.

Next goal is to fix the sensing issues with AAI and VVI then implement egrams.