Adds minor timing fixes to inhibiting modes
This commit adds timing corrective factors for the refractory periods and the pulse widths as well as adds in an activity scaling factor so that we can adjust the activity scale.

There is still weird behaviour when the ppm is -1 the pacemaker rate however I doubt that this can be fixed.

The FRONTENT_CTRL = false command has been removed from the stateflow to further improve timing.