Updates DDD and DDDR stateflow
This commit improves upon the old DDD and DDDR stateflow to be more accurate to the updated understanding of how the refractory periods work.

DDD and DDDR behave as expected when under a continuous atrial and continuous ventricular pace individually however when the heart is beating the same as the pacemaker the ventricle doesn't seem to inhibit even though changes have been made to the stateflow (removing a junction) to ensure that once the VENT_CMP_DETECT is true, it will leave the path of triggering the ventricle.

Next area of improvement is getting egrams data to the DCM.