This commit is because of a simulink glitch
I had to reprogram the ConvertToScale custom C++ code block because it wouldn't open on my simulink. It opened on Mathew's simulink which is really weird but just in case I need to use it in the demo, I've recoded it so that way I can open it up again.

Literally nothing changed.