Adds moving window plotting
This commit has the code structure that allows us to use a deque object to plot a "moving window" of points that match the last 10 seconds of activity.

Right now these points are random points and not egrams data however after this commit I will start integrating this with the DCM such that we can get the actual signal.